+++ packages/client/src/api/dataPlaneComponents.ts
@@ -83,8 +83,48 @@ export type PgRollStatusVariables = {
 } & DataPlaneFetcherExtraProps;
 
 export const pgRollStatus = (variables: PgRollStatusVariables, signal?: AbortSignal) =>
-  dataPlaneFetch<Schemas.PgRollStatusResponse, PgRollStatusError, undefined, {}, {}, PgRollStatusPathParams>({
+  dataPlaneFetch<Schemas.PgRollJobStatusResponse, PgRollStatusError, undefined, {}, {}, PgRollStatusPathParams>({
     url: '/db/{dbBranchName}/pgroll/status',
+    method: 'get',
+    ...variables,
+    signal
+  });
+
+export type PgRollJobStatusPathParams = {
+  /**
+   * The DBBranchName matches the pattern `{db_name}:{branch_name}`.
+   */
+  dbBranchName: Schemas.DBBranchName;
+  /**
+   * The id of the migration job
+   */
+  jobId: Schemas.PgRollMigrationJobID;
+  workspace: string;
+  region: string;
+};
+
+export type PgRollJobStatusError = Fetcher.ErrorWrapper<
+  | {
+      status: 400;
+      payload: Responses.BadRequestError;
+    }
+  | {
+      status: 401;
+      payload: Responses.AuthError;
+    }
+  | {
+      status: 404;
+      payload: Responses.SimpleError;
+    }
+>;
+
+export type PgRollJobStatusVariables = {
+  pathParams: PgRollJobStatusPathParams;
+} & DataPlaneFetcherExtraProps;
+
+export const pgRollJobStatus = (variables: PgRollJobStatusVariables, signal?: AbortSignal) =>
+  dataPlaneFetch<Schemas.PgRollJobStatusResponse, PgRollJobStatusError, undefined, {}, {}, PgRollJobStatusPathParams>({
+    url: '/db/{dbBranchName}/pgroll/jobs/{jobId}',
     method: 'get',
     ...variables,
     signal
@@ -4549,6 +4589,7 @@ export const operationsByTag = {
   branch: {
     applyMigration,
     pgRollStatus,
+    pgRollJobStatus,
     getBranchList,
     getBranchDetails,
     createBranch,