+++ packages/client/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @xata.io/client
 
+## 0.28.2
+
+### Patch Changes
+
+- [#1288](https://github.com/xataio/client-ts/pull/1288) [`c9178e1`](https://github.com/xataio/client-ts/commit/c9178e1e3f2268513e78dcfce396a99a8fca5dfb) Thanks [@SferaDev](https://github.com/SferaDev)! - Update read to use `ReadonlyArray<Identifier>`
+
+## 0.28.1
+
+### Patch Changes
+
+- [#1281](https://github.com/xataio/client-ts/pull/1281) [`9a7e3f5`](https://github.com/xataio/client-ts/commit/9a7e3f5029e53efc6750e9c86bab936427788209) Thanks [@SferaDev](https://github.com/SferaDev)! - Fix issue with default public access
+
 ## 0.28.0
 
 ### Minor Changes