+++ packages/plugin-client-drizzle/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @xata.io/drizzle
 
+## 0.0.13
+
+### Patch Changes
+
+- Updated dependencies [[`c9178e1`](https://github.com/xataio/client-ts/commit/c9178e1e3f2268513e78dcfce396a99a8fca5dfb)]:
+  - @xata.io/client@0.28.2
+
+## 0.0.12
+
+### Patch Changes
+
+- Updated dependencies [[`9a7e3f5`](https://github.com/xataio/client-ts/commit/9a7e3f5029e53efc6750e9c86bab936427788209)]:
+  - @xata.io/client@0.28.1
+
 ## 0.0.11
 
 ### Patch Changes